requirements.txt
alembic==1.0.11
attrs==19.1.0
Click==7.0
Flask==1.0.3
Flask-Caching==1.7.2
Flask-Humanize==0.3.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
geojson==2.4.1
humanfriendly==4.18
humanize==0.5.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
svgwrite==1.3.1
toml==0.10.0
Werkzeug==0.15.4
